+++ config/config.json
@@ -1,0 +1,10 @@
+{
+    "servers_dict": {
+        "A3 TacR": "d7291177",
+        "A3 Training 1": "9fd28d83",
+        "A3 Training 2": "6b20f439",
+        "A3 Training 3": "7b217461",
+        "Sq Training": "7e1451cc",
+        "Sq TacR": "0b493e80"
+    }
+}
